requirements.txt
matplotlib
RPi.GPIO
